+++ src/main/java/org/betonquest/betonquest/events/OpSudoEvent.java
@@ -1,23 +1,31 @@
 package org.betonquest.betonquest.events;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.Instruction;
 import org.betonquest.betonquest.api.QuestEvent;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
 import org.betonquest.betonquest.utils.PlayerConverter;
 import org.bukkit.entity.Player;
 
+import java.util.List;
+
 @SuppressWarnings("PMD.CommentRequired")
 public class OpSudoEvent extends QuestEvent {
 
-    private final String[] commands;
+    private final Command[] commands;
 
     public OpSudoEvent(final Instruction instruction) throws InstructionParseException {
         super(instruction, true);
         final String string = instruction.getInstruction().trim();
         int index = string.indexOf("conditions:");
         index = index == -1 ? string.length() : index;
-        commands = string.substring(string.indexOf(' ') + 1, index).split("\\|");
+        final String[] rawCommands = string.substring(string.indexOf(' ') + 1, index).split("\\|");
+
+        commands = new Command[rawCommands.length];
+        for (int i = 0; i < rawCommands.length; i++) {
+            commands[i] = new Command(rawCommands[i]);
+        }
     }
 
     @Override
@@ -27,8 +35,13 @@ public class OpSudoEvent extends QuestEvent {
         final boolean previousOp = player.isOp();
         try {
             player.setOp(true);
-            for (final String command : commands) {
-                player.performCommand(command.replace("%player%", player.getName()));
+            for (final Command command : commands) {
+                String com = command.command;
+                for (final String var : command.variables) {
+                    com = com.replace(var, BetonQuest.getInstance().getVariableValue(
+                            instruction.getPackage().getName(), var, playerID));
+                }
+                player.performCommand(com);
             }
         } finally {
             player.setOp(previousOp);
@@ -36,4 +49,16 @@ public class OpSudoEvent extends QuestEvent {
         return null;
     }
 
+    @SuppressWarnings("PMD.AvoidFieldNameMatchingTypeName")
+    private static class Command {
+
+        private final String command;
+        private final List<String> variables;
+
+        public Command(final String command) {
+            this.command = command;
+            variables = BetonQuest.resolveVariables(command);
+        }
+
+    }
 }